# perf: optimize bundle strategy and resource loading